requirements_for_dev.txt
pytest==4.2.0
pytest-django==3.4.7
factory-boy==2.11.1
pytest-freezegun==0.3.0
